Update README.md
It is a MongoDB session-based authentication system using the passport framework. I have used express-session and mango store for creating a mongo store, where we can enter our website very easily if we had a previous login or signing history. It will make a cookie that actually works as a piece of proof that we are authenticated. So, as a result, next time we won't need to log in, also I have used bcypt authentication system for generating passwords. And I used ejs to struct and route my pages by making a header and footer...